Postgres was having trouble restablishing connections for non-pool clients.
Postgres' connector-manager was only checking to see if this.client was null. We
should have been checking for "readyForQuery" status as well. We now make
this.client null if the connection has an error rather than just maintaining
stale information. This is to ensure that we retry our connection until we can
re-establish a firm connection.

This resolves issue #832 https://github.com/sequelize/sequelize/issues/832