fix addColumnQuery for ENUM types
There a whole bunch wrong here, but mostly with sequelize. First and
foremost attributeToSQL (the convention) should be working with attributes,
however it's using dataTypes here because.. well every other dialect does
this too. addColumnQuery as defined in the abstract QueryGenerator class
says you should be using attributes, but everyone else uses (key, dataType).
This is a kludge to make ENUMs work for TSQL, since it needs to know the
field name to generate the SQL