Updated checks of error messages to use the new values which are what is propagated from the connection failure.
As pointed out in the [PR](https://github.com/sequelize/sequelize/pull/2576#issuecomment-63018736) for this functionality we would ideally not be checking error messages and instead be checking the constructors of the objects to make sure that they are an instance of the correct error type. There's some issues right now with two different constructors being used and not being able to use `instanceof`. We could fall back to the constructor name or some other check, but at the end of the day that isn't really any better than just checking the message.